Refactor Volvo Master Data Service and ViewModels
- Updated logging method from LogInformationAsync to LogInfoAsync for consistency.
- Replaced Model_Dao_Result with Model_Dao_Result_Factory for error handling in Service_VolvoMasterData.
- Changed DateTime properties to DateTimeOffset in ViewModel_Volvo_History and ViewModel_Volvo_ShipmentEntry for better time zone handling.
- Improved null checks and error handling in various methods across ViewModels.
- Enhanced CSV import/export functionality in ViewModel_Volvo_Settings.
- Added NullableDoubleToDoubleConverter for better binding with NumberBox controls in XAML.
- Cleaned up XAML bindings to use Binding instead of x:Bind where appropriate.
- Removed unnecessary whitespace and improved code readability.

+++ Module_Volvo/Models/Model_VolvoShipmentLine.cs
@@ -45,7 +45,7 @@ public class Model_VolvoShipmentLine
     /// Expected skid count from Volvo packlist (NULL if no discrepancy)
     /// Used to calculate difference: ReceivedSkidCount - ExpectedSkidCount
     /// </summary>
-    public int? ExpectedSkidCount { get; set; }
+    public double? ExpectedSkidCount { get; set; }
 
     /// <summary>
     /// User's note about the discrepancy (NULL if no discrepancy)
